backend/requirements.txt
h11~=0.16.0
pip~=25.0.1
wheel~=0.45.1
requests~=2.32.4
PyYAML~=6.0.2
zipp~=3.23.0
exceptiongroup~=1.3.0
sniffio~=1.3.1
MarkupSafe~=3.0.2
werkzeug~=3.1.3
click~=8.1.8
jinja2~=3.1.6
blinker~=1.9.0
itsdangerous~=2.2.0
httpcore~=1.0.9
idna~=3.10
httpx~=0.28.1
jiter~=0.10.0
numpy~=2.0.2
charset-normalizer~=3.4.2
setuptools~=78.1.0
regex~=2024.11.6
sympy~=1.14.0
mpmath~=1.3.0
torch~=2.7.1
filelock~=3.18.0
tqdm~=4.67.1
scipy~=1.13.1
packaging~=25.0
networkx~=3.2.1
importlib-metadata~=8.7.0
urllib3~=2.5.0
fsspec~=2025.7.0
distro~=1.9.0
openai~=1.96.1
pydantic~=2.11.7
websockets~=15.0.1
anyio~=4.9.0
certifi~=2025.7.14
annotated-types~=0.7.0
tokenizers~=0.21.2
safetensors~=0.5.3
scikit-learn~=1.6.1
transformers~=4.53.2
pillow~=11.3.0
huggingface-hub~=0.33.4
faiss-cpu~=1.11.0.post1
nltk~=3.9.1
fastapi~=0.116.1
starlette~=0.47.1
uvicorn~=0.35.0
flask~=3.1.1
python-dotenv~=1.1.1
sentence-transformers~=5.0.0
protobuf~=5.29.5
PyJWT~=2.10.1